# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for PPDB Application",
        "title": "PPDB API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/school-profiles": {
            "get": {
                "description": "Get a list of all School Profiles.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "School Profiles"
                ],
                "summary": "Get all School Profiles",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.SchoolProfile"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new School Profile. Requires authorization.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "School Profiles"
                ],
                "summary": "Create a new School Profile",
                "parameters": [
                    {
                        "description": "School Profile Request Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SchoolProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/school-profiles/{id}": {
            "get": {
                "description": "Get a School Profile by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "School Profiles"
                ],
                "summary": "Get School Profile by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "School Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.SchoolProfile"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing School Profile by its ID. Requires authorization.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "School Profiles"
                ],
                "summary": "Update School Profile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "School Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "School Profile Request Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SchoolProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a School Profile by its ID. Requires authorization.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "School Profiles"
                ],
                "summary": "Delete School Profile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "School Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.SchoolProfile": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "informasi_lainnya": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "mail": {
                    "type": "string"
                },
                "nama_sekolah": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "system_name": {
                    "type": "string"
                },
                "tahun_berdiri": {
                    "type": "string"
                },
                "tingkatan_sekolah": {
                    "type": "string"
                }
            }
        },
        "models.SchoolProfileRequest": {
            "type": "object",
            "required": [
                "alamat",
                "mail",
                "nama_sekolah",
                "phone",
                "tahun_berdiri",
                "tingkatan_sekolah"
            ],
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "informasi_lainnya": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "mail": {
                    "type": "string",
                    "maxLength": 255
                },
                "nama_sekolah": {
                    "type": "string",
                    "maxLength": 255
                },
                "phone": {
                    "type": "string",
                    "maxLength": 255
                },
                "system_name": {
                    "type": "string",
                    "maxLength": 255
                },
                "tahun_berdiri": {
                    "type": "string",
                    "maxLength": 255
                },
                "tingkatan_sekolah": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}